+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,9 +2,14 @@
 
 Release History
 ===============
-(unreleased)
+
+unreleased
 +++++++++++++++++++
+* `vmss create`: Fixed issue where supplying `--app-gateway ID` would fail.
+* `vm create`: Added `--asgs` support.
 * `vm run-command`: support to run commands on remote VMs
+* `vmss encryption`: (PREVIEW) support vmss disk encryptions
+* `vm perform-maintenance`: support to perform maintenance on a vm
 
 2.0.14 (2017-09-11)
 +++++++++++++++++++